requirements.txt
iniconfig==2.1.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
pytesseract==0.3.13
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
